requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
google-api-python-client==2.151.0
gspread==5.12.4
google-auth==2.36.0
openpyxl==3.1.5
